+++ submodules/TelegramUI/Components/CameraScreen/Sources/CameraScreen.swift
@@ -4848,9 +4848,10 @@ private func sampleBufferFromPixelBuffer(pixelBuffer: CVPixelBuffer) -> CMSample
         return nil
     }
 
-    let attachments: NSArray = CMSampleBufferGetSampleAttachmentsArray(sampleBuffer, createIfNecessary: true)! as NSArray
-    let dict: NSMutableDictionary = attachments[0] as! NSMutableDictionary
-    dict[kCMSampleAttachmentKey_DisplayImmediately as NSString] = true as NSNumber
+    let attachments: NSArray? = CMSampleBufferGetSampleAttachmentsArray(sampleBuffer, createIfNecessary: true) as NSArray?
+    if let dict = attachments?.firstObject as? NSMutableDictionary {
+        dict[kCMSampleAttachmentKey_DisplayImmediately as NSString] = true as NSNumber
+    }
 
     return sampleBuffer
 }