+++ submodules/TelegramUI/Sources/Chat/ChatControllerOpenMessageContextMenu.swift
@@ -178,7 +178,7 @@ extension ChatControllerImpl {
                                     animationCache: animationCache,
                                     animationRenderer: animationRenderer,
                                     isStandalone: false,
-                                    subject: message.areReactionsTags(accountPeerId: self.context.account.peerId) ? .messageTag : .reaction(onlyTop: false),
+                                    subject: message.areReactionsTags(accountPeerId: accountContext.account.peerId) ? .messageTag : .reaction(onlyTop: false),
                                     hasTrending: false,
                                     topReactionItems: reactionItems,
                                     areUnicodeEmojiEnabled: false,